requirements.txt
streamlit==1.45.1
pandas>=2.0.0
plotly>=5.15.0
altair>=5.0.0
earthengine-api>=0.1.400
geemap>=0.29.0
google-auth>=2.20.0
streamlit-option-menu==0.3.6
setuptools
numpy>=1.24.0
scipy